Add init rent scooter page
The rent scooter page is called fragment_selected_scooter.
It is not fully added into the nav_graph yet since I have
trouble inserting the links.
Have also addeded a listener to listen when a google map marker
is clicked. Right now it only gets the name from the marker and
nothing else.